refactor: 모임 시간 조회 응답 수정
